+++ packages/api/src/platforms/vtex/resolvers/mutation.ts
@@ -1,6 +1,5 @@
 import { cancelOrder } from './cancelOrder'
 import { processOrderAuthorization } from './processOrderAuthorization'
-import { setPassword } from './setPassword'
 import { subscribeToNewsletter } from './subscribeToNewsletter'
 import { validateCart } from './validateCart'
 import { validateSession } from './validateSession'
@@ -11,5 +10,4 @@ export const Mutation = {
   subscribeToNewsletter,
   cancelOrder,
   processOrderAuthorization,
-  setPassword,
 }