+++ packages/api/test/integration/schema.test.ts
@@ -84,7 +84,6 @@ const MUTATIONS = [
   'subscribeToNewsletter',
   'cancelOrder',
   'processOrderAuthorization',
-  'setPassword',
 ]
 
 let schema: GraphQLSchema