+++ packages/core/src/sdk/deliveryPromise/useDeliveryPromise.ts
@@ -111,7 +111,9 @@ export function useDeliveryPromise({
 
   const isDeliveryPromiseEnabled = deliveryPromiseConfig.enabled
   const isDynamicEstimateEnabled =
-    deliveryPromiseSettings?.dynamicEstimate?.enabled ?? false
+    deliveryPromiseSettings?.dynamicEstimate?.enabled ?? true
+  const isDeliveryOptionsEnabled =
+    deliveryPromiseSettings?.deliveryOptions?.enabled ?? true
 
   const selectedFacets = useMemo(
     () => selectedFilterFacets ?? searchState.selectedFacets,
@@ -304,10 +306,18 @@ export function useDeliveryPromise({
           key !== DYNAMIC_ESTIMATE_FACET_KEY
       )
     }
+    
+    const filteredFacets = allFacets.filter(({ key }) => {
+      if (!isDeliveryOptionsEnabled && key === DELIVERY_OPTIONS_FACET_KEY) {
+        return false
+      }
 
-    const filteredFacets = !isDynamicEstimateEnabled
-      ? allFacets.filter(({ key }) => key !== DYNAMIC_ESTIMATE_FACET_KEY)
-      : allFacets
+      if (!isDynamicEstimateEnabled && key === DYNAMIC_ESTIMATE_FACET_KEY) {
+        return false
+      }
+
+      return true
+    })
 
     return filteredFacets
       .map((facet) => {