+++ packages/core/src/sdk/search/useFilter.ts
@@ -88,7 +88,10 @@ const reducer = (state: State, action: Action) => {
   return state
 }
 
-export const useFilter = (allFacets: Filter_FacetsFragment[]) => {
+export const useFilter = (
+  allFacets: Filter_FacetsFragment[],
+  initialSelectedFacets?: IStoreSelectedFacet[]
+) => {
   const {
     state: { selectedFacets },
   } = useSearch()
@@ -133,6 +136,16 @@ export const useFilter = (allFacets: Filter_FacetsFragment[]) => {
       }),
     [allFacets, selectedMap]
   )
+
+  // Restore initial PLP facets after clearing filters (e.g. { key: category-n, value: 'electronics' })
+  useEffect(() => {
+    if (initialSelectedFacets && selected.length === 0) {
+      dispatch({
+        type: 'selectFacets',
+        payload: initialSelectedFacets,
+      })
+    }
+  }, [initialSelectedFacets, selected])
 
   useEffect(() => {
     dispatch({