+++ packages/core/test/server/index.test.ts
@@ -89,7 +89,6 @@ const MUTATIONS = [
   'subscribeToNewsletter',
   'cancelOrder',
   'processOrderAuthorization',
-  'setPassword',
 ]
 
 describe('FastStore GraphQL Layer', () => {